Configurations:  'specification/netapp/resource-manager/Microsoft.NetApp/NetApp/tspconfig.yaml', API Version: 2025-09-01, SDK Release Type: stable, and CommitSHA: '756699c964e06e55640096504c33ee4828d027fb' in SpecRepo: 'https://github.com/Azure/azure-rest-api-specs' Pipeline run: https://dev.azure.com/azure-sdk/internal/_build/results?buildId=5561685 Refer to https://eng.ms/docs/products/azure-developer-experience/develop/sdk-release/sdk-release-prerequisites to prepare for SDK release.

+++ sdk/netapp/azure-resourcemanager-netapp/src/main/java/com/azure/resourcemanager/netapp/implementation/CapacityPoolImpl.java
@@ -70,11 +70,11 @@ public final class CapacityPoolImpl implements CapacityPool, CapacityPool.Defini
         return this.innerModel().provisioningState();
     }
 
-    public Double totalThroughputMibps() {
+    public Float totalThroughputMibps() {
         return this.innerModel().totalThroughputMibps();
     }
 
-    public Double utilizedThroughputMibps() {
+    public Float utilizedThroughputMibps() {
         return this.innerModel().utilizedThroughputMibps();
     }
